Make static binaries generation more uniform.
Removes individually made static binaries generation to a for loop
(should remove any errors if / when we do decide to add more
architectures).

Also removes an unused Make target, bundles-ce-binary.tgz didn't seem to
have any use inside of `docker/release` so it was removed. (this from a
quick grep of `bundles/\*/bi.*`)

A bit worried about the switch statement in the static binaries build
step but I don't anticipate it growing any bigger than where it is now.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>