deb/rpm: reuse bundles targets from docker-ce-packaging
This is to avoid duplicate codepaths.
Note that this completes the move of the tar.gz logic from
release-packaging to docker-ce-packaging only for the deb/rpm cases.

Static Linux, Windows and Mac were not touched.

Also I still think this whole bundles tar.gz thing is useless,
but this is what release-repo is currently consuming.

This also makes rpms work, as the path structure changed in docker-ce-packaging
to include distro.

Signed-off-by: Tibor Vass <tibor@docker.com>