Jenkinsfile: set VERIFY_PACKAGE_REPO to prod
Basically this should almost always be prod, even when testing staging.
The only time you may want to set this parameter to stage is if you
have released containerd to staging and want to build docker and test
packages against the newer containerd. It is advised to do a full
containerd release to prod instead.

Signed-off-by: Tibor Vass <tibor@docker.com>