Jenkinsfile: remove RELEASE_STAGING, RELEASE_PRODUCTION
These parameters allowed us to build packages and, once built, trigger
the release-repo and release-verify build pipelines.

In practice, we currently perform these steps separately;

- trigger release-packaging for containerd (creates a `dockereng/containerd-packaging:vXXX` image)
- trigger release-packaging (creates a `jenkins-release-packaging-ce-20.10-XXX` artifact)
- trigger release-repo, and specify which artifacts to publish/release on download.docker.com (or staging)

This patch removes the integration with release-repo/release-verify, so
that this pipeline can only be used to build artifacts, but not to trigger
a full release.
Integration with the "containerd" pipeline was already missing (cannot perform
a containerd build _and_ trigger the rest of the pipeline), so removing this
functionality makes it slightly clearer what the purpose is of this repository;
build artifacts (packages), which can be published to download.docker.com using
the "release-repo" pipeline.

While at it, this also removes some variables that were used to configure the
Slack channels to announce builds on (which were outdated).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>